[修复] 优化应用缓存同步，同时修复应用无法添加删除问题

+++ platform-sms-admin/admin-web/src/main/java/com/courage/platform/sms/admin/api/admin/AppInfoController.java
@@ -51,7 +51,7 @@ public class AppInfoController {
 
     @PostMapping(value = "/deleteAppInfo")
     public ResponseEntity deleteAppInfo(String id) {
-        if (Integer.valueOf(id) == 1) {
+        if (Long.valueOf(id) == 1) {
             logger.info("默认应用不能删除");
             return ResponseEntity.fail("默认应用不能删除");
         }
